added some enemies and a pickup animation
also worked more on temple layout and started on the cave boss

+++ Assets/scripts/inventory/ConsumableClass.cs
@@ -7,6 +7,6 @@ public class ConsumableClass : ItemClass
 
     [Header("consumable")]
     //data specific to consumable class
-    public float healthAdded;
+    public int healthAdded;
     public override ConsumableClass GetConsumable() { return this; }
 }
